1. Adjust the sending logic 2. Only supports sending emails through file queues

+++ Swiftmailer/Spool/EvenlyDistributesSpool.php
@@ -5,6 +5,8 @@ namespace MauticPlugin\MauticEvenlyDistributesSmtpBundle\Swiftmailer\Spool;
 use Mautic\CoreBundle\Helper\CoreParametersHelper;
 use Mautic\EmailBundle\Swiftmailer\Spool\DelegatingSpool;
 use MauticPlugin\MauticEvenlyDistributesSmtpBundle\Swiftmailer\Transport\EvenlyDistributesSmtpTransport;
+use MauticPlugin\MauticEvenlyDistributesSmtpBundle\Helper\CommonHelper;
+use Symfony\Component\DependencyInjection\ContainerInterface as Container;
 
 /**
  * Date: 2024-09-04
@@ -18,6 +20,11 @@ class EvenlyDistributesSpool extends DelegatingSpool
     /** @var EvenlyDistributesSmtpTransport  */
     private $evenlyDistributesSmtpTransport;
 
+    /** @var CommonHelper  */
+    private $commonHelper;
+
+    private $container;
+
     /**
      * DelegatingSpool constructor.
      *
@@ -26,11 +33,15 @@ class EvenlyDistributesSpool extends DelegatingSpool
     public function __construct(
         CoreParametersHelper $coreParametersHelper,
         \Swift_Transport $realTransport,
-        EvenlyDistributesSmtpTransport $evenlyDistributesSmtpTransport
+        EvenlyDistributesSmtpTransport $evenlyDistributesSmtpTransport,
+        CommonHelper $commonHelper,
+        Container $container
     )
     {
         $this->coreParametersHelper = $coreParametersHelper;
         $this->evenlyDistributesSmtpTransport = $evenlyDistributesSmtpTransport;
+        $this->commonHelper = $commonHelper;
+        $this->container = $container;
         parent::__construct($coreParametersHelper, $realTransport);
     }
 
@@ -44,7 +55,7 @@ class EvenlyDistributesSpool extends DelegatingSpool
     public function flushQueue(\Swift_Transport $transport, &$failedRecipients = null)
     {
         $directoryIterator = new \DirectoryIterator($this->getSpoolDir());
-        $transport = $this->evenlyDistributesSmtpTransport;
+//        $transport = $this->evenlyDistributesSmtpTransport;
         /* Start the transport only if there are queued files to send */
         if (!$transport->isStarted()) {
             foreach ($directoryIterator as $file) {
@@ -69,7 +80,17 @@ class EvenlyDistributesSpool extends DelegatingSpool
             if (rename($file, $file.'.sending')) {
                 $message = unserialize(file_get_contents($file.'.sending'));
 
+                $smtp = $this->commonHelper->smtpServersRepository->getActiveServer();
+                if (empty($smtp)) {
+                    echo 'All servers reached 100% daily max';
+                    return 0;
+                }
+                // set transport, message, stat object
+                $this->commonHelper->setTransportAndMessageAndStat($smtp, $transport, $message, $stat);
+
                 $count += $transport->send($message, $failedRecipients);
+                // add successful sent log
+                $this->commonHelper->addSentLog($smtp['id'], $stat);
 
                 unlink($file.'.sending');
             } else {
